requirements.txt
selenium==4.5.0
beautifulsoup4==4.11.1
openai==0.27.0
python-dotenv==0.21.0
pandas==1.5.0
